Add functions to generate UUIDs for msgs, sessions, runs and contacts to ensure consistency
